[FEAT/#3] BottomScreen, Screen으로 분리
